Update map, nearby, and chat tabs
Refresh those tabs so if someone uploads a picture, they appear if you start a chat with them and so you can select their marker fluidly in the map. Also adjust the chat bar so the keyboard doesn't cover it for iPhone, and make it so the chat automatically scrolls down to the latest message after sending one.